requirements.txt
streamlit==1.29.0
pandas==2.1.1
pillow==10.1.0
openpyxl==3.1.2